requirements.txt
Django>1.6.99
django-debug-toolbar==1.2
sqlparse==0.1.11
django-crispy-forms==1.4.0
django-guardian==1.2.0
six==1.6.1
django-allauth==0.16.1
oauthlib==0.6.1
python3-openid==3.0.4
requests==2.3.0
requests-oauthlib==0.4.0
django-filter==0.7
djangorestframework==2.3.14
